zadanie5.cpp
Input:

The input consists of a natural number t, representing the number of test cases. Then t datasets follow. Each dataset contains three natural numbers:

    n: the number of points (nodes),
    m: the number of connections (edges) between them,
    s: the starting point number where the timeline repair begins.

In the following m lines, pairs of natural numbers x and y are provided, indicating that a decision at point x leads to point y. It is guaranteed that:

    There are no duplicate connections,
    There are no self-loops (connections that start and end at the same point).

The points are numbered from 00 to n−1n−1.
Constraints:

    1≤t≤101≤t≤10
    1≤n≤100,0001≤n≤100,000
    0≤m≤2n0≤m≤2n
    0≤x,y<n0≤x,y<n

Output:

For each test case, the program should output:

    NIE if it is impossible for Janusz to stabilize the timeline,
    PETLA if a loop is encountered and stabilization is not possible,
    TAK if Janusz can stabilize the timeline.

If the result is TAK, a second line should include the sequence of point numbers that must be visited in order to achieve stabilization. This sequence must:

    Represent the shortest possible solution,
    Be lexicographically smallest in case of ties.

Example:
Input:

3
6 7 3
0 1
1 5
2 0
2 1
3 2
3 4
4 0
4 3 0
0 1
1 2
2 3
2 2 0
0 1
1 0

Output:

TAK
3 2 0 1 5
NIE
PETLA